pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=6.5.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "python-dotenv>=1.2.1",
]
